requirements.txt
asgiref==3.8.1
cached-property==1.5.2
click==7.1.2
dj-database-url==0.5.0
Django==4.2.16
gunicorn==20.1.0
lxml==4.9.4
packages==0.1.1
psycopg2==2.9.9
pymongo==3.13.0
redis==3.5.3
requests-cache==0.5.2
sqlparse==0.5.1
tqdm==4.66.5
django-summernote-0.8.20.0
